Update test and clean up
