+++ frontends/main/src/page-components/LearningResourceExpanded/LearningResourceExpanded.tsx
@@ -2,8 +2,7 @@ import React, { useEffect, useRef, useState } from "react"
 import styled from "@emotion/styled"
 import { theme } from "ol-components"
 import type { ImageConfig, LearningResourceCardProps } from "ol-components"
-import type { LearningResource } from "api"
-import { ResourceTypeEnum } from "api"
+import type { LearningResource, ResourceTypeEnum } from "api"
 import { useToggle } from "ol-utilities"
 import InfoSection from "./InfoSection"
 import type { User } from "api/hooks/user"
